requirements.txt
streamlit
google-genai
pdf2image
Pillow
gTTS
PyPDF2
pandas
requests
reportlab